DATABASETOOL: started to implement a tool to generate database bindings

+++ src/modules/backend/entity/EntityId.h
@@ -8,6 +8,6 @@
 
 namespace backend {
 
-typedef int32_t EntityId;
+typedef int64_t EntityId;
 
 }
